requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
Pillow==10.1.0
django-cors-headers==4.3.0
django-filter==23.3
drf-yasg==1.21.7  # For API documentation
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
django-storages==1.14.2
boto3==1.28.64  # For AWS S3 file storage
django-ratelimit==3.0.1
django-celery-beat==2.5.0
firebase-admin>=6.0.0
psycopg2-binary==2.9.9